refactor(heureka): remove re-export through index.ts
