+++ apps/heureka/src/components/ServiceDetails/index.tsx
@@ -7,7 +7,7 @@ import React, { useState } from "react"
 import { Stack, Pill, DataGrid, DataGridRow, DataGridHeadCell, DataGridCell } from "@cloudoperators/juno-ui-components"
 import { MessagesProvider, Messages } from "@cloudoperators/juno-messages-provider"
 import { ServiceImageVersions } from "../common/ServiceImageVersions"
-import { ImageVersionDetailsPanel } from "./ImageVersionDetailsPanel/ImageVersionDetailsPanel"
+import { ImageVersionDetailsPanel } from "./ImageVersionDetailsPanel"
 import { useStore } from "../../store/StoreProvider"
 import { SelectServiceDetailsPayload, UserView } from "../../store/StoreProvider/types"
 import { Breadcrumb } from "../common/Breadcrumb"
