--- apps/heureka/src/components/common/ErrorBoundary/index.ts
@@ -1,6 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2024 SAP SE or an SAP affiliate company and Juno contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export { ErrorBoundary } from "./ErrorBoundary"
